adding reports module
